Move checkout to its own controller - add checkoutController.php - remove from bookingController - amend routes - update links
